# public/resources/swagger/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Préstamos",
        "description": "Documentación generada automáticamente con OpenAPI/Swagger para la API de préstamos.",
        "contact": {
            "name": "API prestamos",
            "email": "isaaccaliz305@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/solicitudes": {
            "get": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Obtener lista de solicitudes",
                "description": "Devuelve todas las solicitudes de préstamo registradas (requiere autenticación).",
                "operationId": "07656bdd4c54e32b2596980ff36ce6c0",
                "responses": {
                    "200": {
                        "description": "Lista de solicitudes obtenida correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Solicitud"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error interno del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Crear nueva solicitud",
                "description": "Registra una nueva solicitud de préstamo para un cliente (requiere autenticación).",
                "operationId": "07c5e9e87e8656e1a46ae55e7e5a727a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cliente_id",
                                    "monto_solicitado",
                                    "plazo_meses",
                                    "tasa_interes"
                                ],
                                "properties": {
                                    "cliente_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "monto_solicitado": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 15000.5
                                    },
                                    "plazo_meses": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "tasa_interes": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 12.5
                                    },
                                    "observaciones": {
                                        "type": "string",
                                        "example": "Cliente solicita préstamo para vehículo"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Solicitud creada correctamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/{id}": {
            "get": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Obtener solicitud específica",
                "description": "Obtiene los detalles de una solicitud de préstamo específica (requiere autenticación).",
                "operationId": "d7fd9e42643919feb14c753698ca75b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a consultar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud encontrada"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Actualizar solicitud",
                "description": "Modifica una solicitud existente (por ejemplo, aprobar o rechazar).",
                "operationId": "3a9425b677c6d8c04a9dcee0a60acf00",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud actualizada correctamente"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Eliminar solicitud",
                "description": "Elimina una solicitud si aún no tiene préstamo aprobado (requiere autenticación).",
                "operationId": "0c67aa5acf3f079065fb170e0b3de8ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Solicitud eliminada correctamente"
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/{id}/aprobar": {
            "put": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Aprobar solicitud",
                "description": "Cambia el estado de una solicitud a APROBADO.",
                "operationId": "20d6be21acfaa0cfe8c47d43390231b7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a aprobar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud aprobada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud aprobada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    },
                    "400": {
                        "description": "Solicitud ya estaba aprobada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/{id}/rechazar": {
            "put": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Rechazar solicitud",
                "description": "Cambia el estado de una solicitud a RECHAZADO.",
                "operationId": "253c0959b3e16fa22d33a4bcab90f87e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la solicitud a rechazar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Solicitud rechazada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Solicitud rechazada correctamente."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Solicitud no encontrada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/solicitudes/cliente/{cliente_id}": {
            "get": {
                "tags": [
                    "Solicitudes"
                ],
                "summary": "Listar solicitudes por cliente",
                "description": "Obtiene todas las solicitudes registradas de un cliente específico.",
                "operationId": "5758b888167eb7777dfa0aa277b0d581",
                "parameters": [
                    {
                        "name": "cliente_id",
                        "in": "path",
                        "description": "ID del cliente",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Listado de solicitudes obtenido correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Solicitud"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Cliente": {
                "title": "Cliente",
                "description": "Modelo que representa un cliente en el sistema",
                "required": [
                    "id",
                    "primer_nombre",
                    "primer_apellido",
                    "dpi",
                    "nit"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "primer_nombre": {
                        "type": "string",
                        "example": "Juan"
                    },
                    "segundo_nombre": {
                        "type": "string",
                        "example": "Carlos"
                    },
                    "primer_apellido": {
                        "type": "string",
                        "example": "Pérez"
                    },
                    "segundo_apellido": {
                        "type": "string",
                        "example": "López"
                    },
                    "dpi": {
                        "type": "string",
                        "example": "1234567890101"
                    },
                    "nit": {
                        "type": "string",
                        "example": "1234567-8"
                    },
                    "fecha_nacimiento": {
                        "type": "string",
                        "format": "date",
                        "example": "1990-05-10"
                    },
                    "direccion": {
                        "type": "string",
                        "example": "Ciudad de Guatemala, Zona 1"
                    },
                    "correo": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "telefono": {
                        "type": "string",
                        "example": "5555-1234"
                    }
                },
                "type": "object"
            },
...
            "Solicitud": {
                "required": [
                    "cliente_id",
                    "monto_solicitado",
                    "plazo_meses",
                    "tasa_interes"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "cliente_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "monto_solicitado": {
                        "type": "number",
                        "format": "float",
                        "example": 15000.5
                    },
                    "plazo_meses": {
                        "type": "integer",
                        "example": 12
                    },
                    "tasa_interes": {
                        "type": "number",
                        "format": "float",
                        "example": 12.5
                    },
                    "estado": {
                        "type": "string",
                        "example": "EN PROCESO"
                    },
                    "observaciones": {
                        "type": "string",
                        "example": "Solicitud para compra de vehículo."
                    },
                    "usuario_crea": {
                        "type": "string",
                        "example": "admin"
                    },
                    "usuario_actualiza": {
                        "type": "string",
                        "example": "operador1"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "cliente": {
                        "$ref": "#/components/schemas/Cliente"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}